pyproject.toml
[project]
name = "g2"
version = "1.0.0"
description = "Personal Claude assistant. Lightweight, secure, customizable."
requires-python = ">=3.12"
dependencies = [
    "neonize",
    "aiosqlite>=0.20",
    "croniter>=2.0",
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "structlog>=24.0",
    "qrcode[pil]",
    "pydantic>=2.0",
    "watchfiles>=0.21",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
g2 = "g2.__main__:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/g2", "skills_engine"]

# --- uv ---

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov",
    "ruff>=0.5",
    "mypy>=1.10",
]

# --- ruff ---

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests", "skills_engine"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["g2", "skills_engine"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# --- pytest ---

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

# --- mypy ---

[tool.mypy]
python_version = "3.12"
strict = true
